Accept a processor function in `unpackLOD`
Instead of binding the options to `unpackDEF`, that is `unpackPCX`. This gives much more freedom when parsing files, including special treatment of other file formats.

Alternatively, to use specific processors for specific file formats, an object may be supplied, whose keys are case insensitively matched against the file extension, and the corresponding processor triggered.

Processors are supplied with the `ArrayBuffer` and name of the parsing file, and it's return value stored.