Add `.transparency` to `unpackPCX`
It checks if the first palette entry of a PCX matches any of the supplied colours, and if so, replaces that colour with transparent pixels.

It defaults to the colours that the game makes transparent: [{r:0,g:255,b:255}, {r:255,g:0,b:255}, {r:0,g:0,b:0}].

Note that only references to the recognised colour at index 0 will work; in other words, a PCX can use the same colour for transparency and for image data.